* 2620_cleanup: (25 commits)
  vfs_parse_ls_lga: The checks for timestamp is commented out.
  Include config.h to all test's files
  vfs_parse_ls_lga: Make tests output more verbose
  Ticket #2627: Fixed several errors in manpages.
  Fix build with --disable-charset option.
  Added the 'c' hotkey for 'Compile and link current .c file' item of user menu.
  Editor: reduce scope of some functions.
  Editor: reduce scope of some functions.
  Editor: reduce scope of some functions.
  (get_paragraph): fix of pointer difference.
  (load_keymap_from_section): minor optimization.
  (load_keymap_from_section): remove unneeded keybind conversions.
  Added percent sign to key names.
  keymap files: unification of Fxx keys: move to lower case.
  (mc_refresh): moved from lib/widget/wtools.[ch] to lib/widget/dialog-switch.[ch]
  Man pages clean up
  Remove CK_Help action handler in diffviewer
  (size_trunc_len): fixed potential integer overflow if SI is used.
  VFS small optimization
  Remove extra screen update after show help.
  ...

+++ src/editor/wordproc.c
@@ -184,7 +184,8 @@ get_paragraph (WEdit * edit, long p, long q, int indent, int *size)
                     p++;
         *s = edit_get_byte (edit, p);
     }
-    *size = (unsigned long) s - (unsigned long) t;
+    *size = (unsigned long) (s - t);
+    /* FIXME: all variables related to 'size' should be fixed */
     t[*size] = '\n';
     return t;
 }
